+++ packages/plugin-unsupported-blocks/web/src/defaults.ts
@@ -1,0 +1,8 @@
+import { UNSUPPORTED_BLOCKS_TYPE as type } from 'wix-rich-content-plugin-commons';
+export const DEFAULTS = Object.freeze({
+  type,
+  config: {
+    size: 'content',
+    alignment: 'center',
+  },
+});