+++ packages/plugin-unsupported-blocks/web/src/types.ts
@@ -1,0 +1,4 @@
+import { EditorPluginConfig } from 'wix-rich-content-common';
+export interface UnsupportedBlocksPluginEditorConfig extends EditorPluginConfig {
+  [key: string]: any; // eslint-disable-line @typescript-eslint/no-explicit-any
+}